finance/austrian-tax-advisor/chatbot/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.40.0
python-dotenv>=1.0.0
